requirements.txt
django
openai
skidl
python-dotenv
